+++ AddOns/Details/core/parser.lua
@@ -190,12 +190,12 @@
 		---@field key7 number
 
 		local augmentation_aura_list = {
-			[395152] = true,--ebon might (evoker 10.1.5) 395296 = the evoker buff on it self
-			[413984] = true,
-			[410089] = true,--prescience (evoker 10.1.5)
-			[409560] = true,
-			[360827] = true,
-			[410263] = true,
+			[395152] = true, --ebon might (evoker 10.1.5) 395296 = the evoker buff on it self
+			[413984] = true, --Shifting Sands
+			[410089] = true, --prescience (evoker 10.1.5)
+			[409560] = true, --Temporal Wound
+			[360827] = true, --Blistering Scales
+			[410263] = true, --Inferno's Blessing
 		}
 
 		--list of buffs given by another player but should also be credited to the which received it
@@ -213,6 +213,7 @@
 		local augmentation_cache = {
 			ebon_might = {},
 			prescience = {},
+			prescience_stacks = {},
 			---@type table<serial, evokereonsbreathinfo[]>
 			breath_targets = {},
 			flyaway = {},
@@ -455,9 +456,6 @@
 	--tbc spell caches
 	local TBC_PrayerOfMendingCache = {}
 	local TBC_EarthShieldCache = {}
-	local TBC_JudgementOfLightCache = {
-		_damageCache = {}
-	}
 
 	--expose the override spells table to external scripts
 	Details.OverridedSpellIds = override_spellId
@@ -478,6 +476,9 @@
 
 		--Ozumat - Throne of Tides
 		[44566] = true,
+
+		--Smoldering Seedling trinket
+		[212590] = true,
 	}
 
 	local ignored_npcids = {}
@@ -781,7 +782,6 @@
 					sourceFlags = reflection.who_flags
 
 					--data of the aura that caused the reflection
-					--print("2", spellid, GetSpellInfo(spellid))
 					isreflected = spellId --which spell was reflected
 					spellId = reflection.spellid --which spell made the reflection
 					spellName = reflection.spellname
@@ -886,15 +886,6 @@
 			end
 		end
 
-		--wrath of the lich king
-		if (isWOTLK) then
-			--is the target an enemy with judgement of light?
-			if (TBC_JudgementOfLightCache[targetName] and false) then
-				--store the player name which just landed a damage
-				TBC_JudgementOfLightCache._damageCache[sourceName] = {time, targetName}
-			end
-		end
-
 	------------------------------------------------------------------------------------------------
 		--check if need start an combat
 		if (not _in_combat) then --~startcombat ~combatstart
@@ -920,7 +911,7 @@
 
 					Details.WhoAggroTimer = C_Timer.NewTimer(0.1, whoAggro)
 					Details.WhoAggroTimer.HitBy = "|cFFFFFF00First Hit|r: " .. (link or "") .. " from " .. (sourceName or "Unknown")
-					print("debug:", Details.WhoAggroTimer.HitBy)
+					Details:Msg("(debug):", Details.WhoAggroTimer.HitBy)
 				end
 
 				Details:EntrarEmCombate(sourceSerial, sourceName, sourceFlags, targetSerial, targetName, targetFlags)
@@ -1354,6 +1345,8 @@
 						evokerActor.total_extra = evokerActor.total_extra + predictedAmount
 						augmentedSpell.total = augmentedSpell.total + predictedAmount
 						augmentedSpell.targets[sourceName] = (augmentedSpell.targets[sourceName] or 0) + predictedAmount
+
+						--DetailsParserDebugFrame:BlinkIcon(extraSpellId, 1)
 					end
 				end
 			end
@@ -1392,6 +1385,8 @@
 
 						augmentedSpell.total = augmentedSpell.total + predictedAmount
 						augmentedSpell.targets[sourceName] = (augmentedSpell.targets[sourceName] or 0) + predictedAmount
+
+						--DetailsParserDebugFrame:BlinkIcon(extraSpellId, 2)
 					end
 				end
 			end
@@ -1427,6 +1422,8 @@
 
 						augmentedSpell.total = augmentedSpell.total + damageSplitted
 						augmentedSpell.targets[sourceName] = (augmentedSpell.targets[sourceName] or 0) + damageSplitted
+
+						--DetailsParserDebugFrame:BlinkIcon(extraSpellId, 3)
 					end
 				end
 			end
@@ -1462,6 +1459,8 @@
 
 						augmentedSpell.total = augmentedSpell.total + fateMirror_plus_Prescience
 						augmentedSpell.targets[sourceName] = (augmentedSpell.targets[sourceName] or 0) + fateMirror_plus_Prescience
+
+						--DetailsParserDebugFrame:BlinkIcon(extraSpellId, 4)
 					end
 				end
 			end
@@ -1495,6 +1494,8 @@
 
 						augmentedSpell.total = augmentedSpell.total + amount
 						augmentedSpell.targets[sourceName] = (augmentedSpell.targets[sourceName] or 0) + amount
+
+						--DetailsParserDebugFrame:BlinkIcon(extraSpellId, 5)
 					end
 				end
 			end
@@ -1726,7 +1727,7 @@
 		local this_event = t[i]
 
 		if (not this_event) then
-			return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+			return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 		end
 
 		this_event [1] = true --true if this is a damage || false for healing
@@ -1821,7 +1822,7 @@
 		local this_event = t [i]
 
 		if (not this_event) then
-			return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+			return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 		end
 
 		this_event [1] = true --true if this is a damage || false for healing
@@ -1941,7 +1942,7 @@
 		local this_event = t [i]
 
 		if (not this_event) then
-			return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+			return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 		end
 
 		this_event [1] = true --true if this is a damage || false for healing
@@ -2416,6 +2417,19 @@
 			return
 		end
 
+		--> npcId check for ignored npcs
+		--> get the npcId from the cache, if it's not there then get it from the serial and add it to the cache
+		local npcId = npcid_cache[targetSerial] --target npc
+		if (not npcId) then
+			--this string manipulation is running on every event
+			npcId = tonumber(select(6, strsplit("-", targetSerial)) or 0)
+			npcid_cache[targetSerial] = npcId
+		end
+
+		if (ignored_npcids[npcId]) then
+			return
+		end
+
 		if (not sourceName) then
 			--no actor name, use spell name instead
 			sourceName = names_cache[spellName]
@@ -2466,24 +2480,9 @@
 					TBC_PrayerOfMendingCache[sourceName] = nil
 				end
 
-			elseif (spellId == 27163 and false) then --Judgement of Light (paladin) --disabled on 25 September 2022
-				--check if the hit was landed in the same cleu tick
-
-				local hitCache = TBC_JudgementOfLightCache._damageCache[sourceName]
-				if (hitCache) then
-					local timeLanded = hitCache[1]
-					local targetHit = hitCache[2]
-
-					if (timeLanded and timeLanded == time) then
-						local sourceData = TBC_JudgementOfLightCache[targetHit]
-						if (sourceData) then
-							--change the source of the healing
-							sourceSerial, sourceName, sourceFlags = unpack(sourceData)
-							--erase the hit time information
-							TBC_JudgementOfLightCache._damageCache[sourceName] = nil
-						end
-					end
-				end
+			elseif (spellId == 27163) then --Judgement of Light (paladin), reattribute healing to the person getting healed. Stops 'sniping' the JoL to parse
+				--Removed old version 10/27/23 Flamanis
+				sourceSerial, sourceName, sourceFlags = targetSerial, targetName, targetFlags
 			end
 		end
 
@@ -2854,7 +2853,7 @@
 				local thisEvent = deathLog[i]
 
 				if (not thisEvent) then
-					return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+					return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 				end
 
 				thisEvent[1] = 5 --5 = buff aplication
@@ -2917,9 +2916,6 @@
 
 				elseif (spellId == SPELLID_PRIEST_POM_BUFF) then
 					TBC_PrayerOfMendingCache [targetName] = {sourceSerial, sourceName, sourceFlags}
-
-				elseif (spellId == 27163 and false) then --Judgement Of Light
-					TBC_JudgementOfLightCache[targetName] = {sourceSerial, sourceName, sourceFlags}
 				end
 			end
 
@@ -2961,12 +2957,6 @@
 	--recording debuffs applied by player
 
 		elseif (auraType == "DEBUFF") then
-			if (isWOTLK) then --buff applied
-				if (spellId == 27162 and false) then --Judgement Of Light
-					--which player applied the judgement of light on this mob
-					TBC_JudgementOfLightCache[targetName] = {sourceSerial, sourceName, sourceFlags}
-				end
-			end
 
 		------------------------------------------------------------------------------------------------
 		--spell reflection
@@ -3074,12 +3064,6 @@
 	--recording debuffs applied by player
 
 		elseif (tipo == "DEBUFF") then
-			if (isWOTLK) then --buff refresh
-				if (spellId == 27162 and false) then --Judgement Of Light
-					--which player applied the judgement of light on this mob
-					TBC_JudgementOfLightCache[targetName] = {sourceSerial, sourceName, sourceFlags}
-				end
-			end
 
 			if (_in_combat) then
 				------------------------------------------------------------------------------------------------
@@ -3173,11 +3157,6 @@
 	------------------------------------------------------------------------------------------------
 	--recording debuffs applied by player
 		elseif (tipo == "DEBUFF") then
-			if (isWOTLK) then --buff removed
-				if (spellId == 27162 and false) then --Judgement Of Light
-					TBC_JudgementOfLightCache[targetName] = nil
-				end
-			end
 
 		------------------------------------------------------------------------------------------------
 		--spell reflection
@@ -3375,7 +3354,7 @@
 					local this_event = t [i]
 
 					if (not this_event) then
-						return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+						return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 					end
 
 					this_event [1] = 4 --4 = debuff aplication
@@ -3422,7 +3401,7 @@
 					local this_event = t [i]
 
 					if (not this_event) then
-						return print("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
+						return Details:Msg("Parser Event Error -> Set to 16 DeathLogs and /reload", i, _amount_of_last_events)
 					end
 
 					this_event [1] = 4 --4 = debuff aplication
@@ -3978,7 +3957,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			petName, ownerName, ownerGUID, ownerFlags = Details.tabela_pets:GetPetOwner(sourceSerial, sourceName, sourceFlags)
 			if (petName) then
 				ownerActor = _current_misc_container:GetOrCreateActor(ownerGUID, ownerName, ownerFlags, true)
-				--print("pet found:", petName, ownerName, ownerGUID, ownerFlags)
 			end
 		end
 
@@ -4130,10 +4108,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		local amountOfCasts = _current_combat.amountCasts[sourceName][spellName] or 0
 		amountOfCasts = amountOfCasts + 1
 		_current_combat.amountCasts[sourceName][spellName] = amountOfCasts
-
-		--if (sourceSerial == UnitGUID("player")) then
-		--	print(sourceName, spellName, amountOfCasts)
-		--end
 
 	------------------------------------------------------------------------------------------------
 	--record cooldowns cast which can't track with buff applyed
@@ -4656,7 +4630,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 								newEventTable[4] = enemyCastTime --when the event happened using unix time
 								newEventTable[5] = 0 --player health when the event happened
 								newEventTable[6] = enemyName --source name
-								--print("addin enemy cast event", alvo_name, i, enemyCastTime+0.1, ">", eventTime)
 								tinsert(eventsBeforePlayerDeath, i, newEventTable)
 								currentEnemyCastIndex = enemyCastEventIndex + 1
 								break
@@ -5184,6 +5157,29 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		return Details.zone_type
 	end
 
+	local gotAggro = false
+	function Details.parser_functions:UNIT_FLAGS(...)
+		if (gotAggro) then
+			return
+		end
+
+		if (Details:GetZoneType() ~= "raid" and Details:GetZoneType() ~= "party") then
+			return
+		end
+
+		local unitId = ...
+
+		if (UnitExists(unitId)) then
+			if (UnitAffectingCombat(unitId) and not UnitAffectingCombat("player")) then
+				Details.LastAggro = UnitName(unitId)
+				gotAggro = true
+				C_Timer.After(1, function()
+					gotAggro = false
+				end)
+			end
+		end
+	end
+
 	function Details.parser_functions:ZONE_CHANGED_NEW_AREA(...)
 		return Details.Schedules.After(1, Details.Check_ZONE_CHANGED_NEW_AREA)
 	end
@@ -5199,6 +5195,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		_in_resting_zone = IsResting()
 
 		parser:WipeSourceCache()
+		Details.listener:UnregisterEvent("UNIT_FLAGS")
 
 		_is_in_instance = false
 
@@ -5220,10 +5217,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		end
 
 		Details.time_type = Details.time_type_original
-
-		if (Details.debug) then
-			Details:Msg("(debug) zone change:", Details.zone_name, "is a", Details.zone_type, "zone.")
-		end
 
 		if (Details.is_in_arena and zoneType ~= "arena") then
 			Details:LeftArena()
@@ -5307,6 +5300,8 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 						Details.ScheduleLoadStorage()
 					end
 				end
+
+				Details.listener:RegisterEvent("UNIT_FLAGS")
 			end
 
 			if (Details:IsInInstance()) then
@@ -5353,7 +5348,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		if (zoneType == "party") then
 			local openRaidLib = LibStub:GetLibrary("LibOpenRaid-1.0", true)
 			if (openRaidLib) then
-				print("sent my keystone to party")
 				openRaidLib.KeystoneInfoManager.SendPlayerKeystoneInfoToParty()
 			end
 		end
@@ -5361,6 +5355,8 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		if (Details.InstancesToStoreData[zoneMapID]) then
 			Details.current_exp_raid_encounters[encounterID] = true
 		end
+
+		Details222.DebugMsg("|cFFFFFF00Who Aggro by UNIT_FLAGS:", Details.LastAggro)
 
 		if (not Details.WhoAggroTimer and Details.announce_firsthit.enabled) then
 			Details.WhoAggroTimer = C_Timer.NewTimer(0.1, whoAggro)
@@ -5429,8 +5425,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			return
 		end
 
-		Details222.Cache.ClearAugmentationCache()
-
 		Details.latest_ENCOUNTER_END = Details.latest_ENCOUNTER_END or 0
 		if (Details.latest_ENCOUNTER_END + 15 > GetTime()) then
 			return
@@ -5478,6 +5472,8 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		end
 
 		Details:SendEvent("COMBAT_ENCOUNTER_END", nil, ...)
+
+		Details222.Cache.ClearAugmentationCache()
 
 		Details:Destroy(Details.encounter_table)
 		Details:Destroy(dk_pets_cache.army)
@@ -5666,12 +5662,9 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 	--can also run when the player leaves combat state (regen enabled)
 	function Details:RunScheduledEventsAfterCombat(OnRegenEnabled)
 		if (Details.debug) then
-			Details:Msg("(debug) running scheduled events after combat end.")
+			--Details:Msg("(debug) running scheduled events after combat end.")
 		end
 
-		TBC_JudgementOfLightCache = {
-			_damageCache = {}
-		}
 
 		--when the user requested data from the storage but is in combat lockdown
 		if (Details.schedule_storage_load) then
@@ -5774,20 +5767,14 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		wipe(keystoneLevels)
 		local libOpenRaid = LibStub("LibOpenRaid-1.0", true)
 
-		--print("saveGroupMembersKeystoneLevel() called", libOpenRaid, GetNumGroupMembers()-1)
-
 		for i = 1, GetNumGroupMembers()-1 do
 			local unitId = "party" .. i
 			if (UnitExists(unitId)) then
 				local unitKeystoneInfo = libOpenRaid.GetKeystoneInfo(unitId)
-				--print("unitExists", unitId, unitKeystoneInfo)
 				if (unitKeystoneInfo) then
 					local unitName = Details:GetFullName(unitId)
 					keystoneLevels[unitName] = unitKeystoneInfo.level
-					--print("saved keystone level for", unitName, unitKeystoneInfo.level)
 				end
-			else
-				--print("unit does not exist", unitId)
 			end
 		end
 
@@ -5797,7 +5784,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			if (unitKeystoneInfo) then
 				local unitName = Details:GetFullName(unitId)
 				keystoneLevels[unitName] = unitKeystoneInfo.level
-				--print("saved keystone level for", unitName, unitKeystoneInfo.level)
 			end
 		end
 	end
@@ -5855,9 +5841,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			C_Timer.After(0, function()
 				if (ChallengeModeCompleteBanner) then
 					ChallengeModeCompleteBanner.timeToHold = 0.1
-					--print("ChallengeModeCompleteBanner.timeToHold Existed!")
-				else
-					--print("ChallengeModeCompleteBanner.timeToHold DID NOT Existed!")
 				end
 			end)
 		end
@@ -5902,10 +5885,6 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 	end
 
 	function Details.parser_functions:PLAYER_REGEN_ENABLED(...)
-		if (Details.debug) then
-			Details:Msg("(debug) |cFFFFFF00PLAYER_REGEN_ENABLED|r event triggered.")
-		end
-
 		if (Details.auto_swap_to_dynamic_overall) then
 			Details:InstanceCall(autoSwapDynamicOverallData, false)
 		end
@@ -5999,6 +5978,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		end
 	end
 
+	--[=[
 	--this is mostly triggered when the player enters in a dual against another player
 	function Details.parser_functions:UNIT_FACTION(unit)
 		if (true) then
@@ -6051,6 +6031,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			end
 		end
 	end
+	--]=]
 
 	function Details.parser_functions:ROLE_CHANGED_INFORM(...)
 		if (Details.last_assigned_role ~= _UnitGroupRolesAssigned("player")) then
@@ -6687,39 +6668,39 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		for n, nn in pairs(damage_cache) do
 			amount = amount + 1
 		end
-		print("parser damage_cache", amount)
+		Details:Msg("parser damage_cache", amount)
 
 		amount = 0
 		for n, nn in pairs(damage_cache_pets) do
 			amount = amount + 1
 		end
-		print("parser damage_cache_pets", amount)
+		Details:Msg("parser damage_cache_pets", amount)
 
 		amount = 0
 		for n, nn in pairs(damage_cache_petsOwners) do
 			amount = amount + 1
 		end
-		print("parser damage_cache_petsOwners", amount)
+		Details:Msg("parser damage_cache_petsOwners", amount)
 
 		amount = 0
 		for n, nn in pairs(healing_cache) do
 			amount = amount + 1
 		end
-		print("parser healing_cache", amount)
+		Details:Msg("parser healing_cache", amount)
 
 		amount = 0
 		for n, nn in pairs(energy_cache) do
 			amount = amount + 1
 		end
-		print("parser energy_cache", amount)
+		Details:Msg("parser energy_cache", amount)
 
 		amount = 0
 		for n, nn in pairs(misc_cache) do
 			amount = amount + 1
 		end
-		print("parser misc_cache", amount)
-		print("group damage", #Details.cache_damage_group)
-		print("group damage", #Details.cache_healing_group)
+		Details:Msg("parser misc_cache", amount)
+		Details:Msg("group damage", #Details.cache_damage_group)
+		Details:Msg("group damage", #Details.cache_healing_group)
 	end
 
 	function Details:GetActorsOnDamageCache()
@@ -6730,14 +6711,17 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		return Details.cache_healing_group
 	end
 
+	--called when the zone type changes and ENCOUNTER_END event
 	function Details222.Cache.ClearAugmentationCache()
 		Details:Destroy(augmentation_cache.ebon_might) --~roskash
 		Details:Destroy(augmentation_cache.prescience)
 		Details:Destroy(augmentation_cache.shield)
 		Details:Destroy(augmentation_cache.infernobless)
 		Details:Destroy(augmentation_cache.breath_targets)
+		Details:Destroy(augmentation_cache.prescience_stacks)
 	end
 
+	--called when restaring the garbage collector, on some options change, at the end of a combat (before COMBAT_PLAYER_LEAVE and after COMBAT_PLAYER_LEAVING)
 	function Details:ClearParserCache(bIsFromCombatStart) --~wipe
 		Details:Destroy(damage_cache)
 		Details:Destroy(damage_cache_pets)
@@ -6770,6 +6754,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 			if (Details.zone_type ~= "party") then
 				Details:Destroy(augmentation_cache.ebon_might) --~roskash
 				Details:Destroy(augmentation_cache.prescience)
+				Details:Destroy(augmentation_cache.prescience_stacks)
 				Details:Destroy(augmentation_cache.shield)
 				Details:Destroy(augmentation_cache.infernobless)
 			end
@@ -7083,7 +7068,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 
 		is_using_spellId_override = Details.override_spellids
 
-		return Details:ClearParserCache(bIsFromCombatStart)
+		Details:ClearParserCache(bIsFromCombatStart)
 	end
 
 	function Details.DumpIgnoredNpcs()
@@ -7248,7 +7233,7 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 
 		for i = 1, players do
 			local name, killingBlows, honorableKills, deaths, honorGained, faction, race, rank, class, classToken, damageDone, healingDone, bgRating, ratingChange, preMatchMMR, mmrChange, talentSpec
-			if (isWOTLK) then
+			if (isWOTLK or isERA) then
 				name, killingBlows, honorableKills, deaths, honorGained, faction, rank, race, class, classToken, damageDone, healingDone, bgRating, ratingChange, preMatchMMR, mmrChange, talentSpec = GetBattlefieldScore(i)
 			else
 				name, killingBlows, honorableKills, deaths, honorGained, faction, race, class, classToken, damageDone, healingDone, bgRating, ratingChange, preMatchMMR, mmrChange, talentSpec = GetBattlefieldScore(i)
@@ -7324,4 +7309,53 @@ local SPELL_POWER_PAIN = SPELL_POWER_PAIN or (PowerEnum and PowerEnum.Pain) or 1
 		end
 	end
 
+--[=[
+local detailsParserDebugFrame = CreateFrame("frame", "DetailsParserDebugFrame", UIParent)
+detailsParserDebugFrame:SetSize(100, 200)
+DetailsFramework:ApplyStandardBackdrop(detailsParserDebugFrame)
+detailsParserDebugFrame:SetPoint("left", UIParent, "left", 2, 350)
+detailsParserDebugFrame.AllIcons = {}
+--detailsParserDebugFrame:Hide()
+local iconSize = 40
 
+local spellIcon1 = detailsParserDebugFrame:CreateTexture(nil, "overlay")
+spellIcon1:SetSize(iconSize, iconSize)
+spellIcon1:SetPoint("topleft", detailsParserDebugFrame, "topleft", 2, -5)
+
+local spellIcon2 = detailsParserDebugFrame:CreateTexture(nil, "overlay")
+spellIcon2:SetSize(iconSize, iconSize)
+spellIcon2:SetPoint("topleft", spellIcon1, "bottomleft", 0, -2)
+
+local spellIcon3 = detailsParserDebugFrame:CreateTexture(nil, "overlay")
+spellIcon3:SetSize(iconSize, iconSize)
+spellIcon3:SetPoint("topleft", spellIcon2, "bottomleft", 0, -2)
+
+local spellIcon4 = detailsParserDebugFrame:CreateTexture(nil, "overlay")
+spellIcon4:SetSize(iconSize, iconSize)
+spellIcon4:SetPoint("topleft", spellIcon3, "bottomleft", 0, -2)
+
+local spellIcon5 = detailsParserDebugFrame:CreateTexture(nil, "overlay")
+spellIcon5:SetSize(iconSize, iconSize)
+spellIcon5:SetPoint("topleft", spellIcon4, "bottomleft", 0, -2)
+
+tinsert(detailsParserDebugFrame.AllIcons, spellIcon1)
+tinsert(detailsParserDebugFrame.AllIcons, spellIcon2)
+tinsert(detailsParserDebugFrame.AllIcons, spellIcon3)
+tinsert(detailsParserDebugFrame.AllIcons, spellIcon4)
+tinsert(detailsParserDebugFrame.AllIcons, spellIcon5)
+
+function detailsParserDebugFrame:BlinkIcon(spellId, iconId)
+	local spellName, _, spellIcon = GetSpellInfo(spellId)
+	local icon = self.AllIcons[iconId]
+
+	if (spellIcon) then
+		icon:SetTexture(spellIcon)
+		icon:Show()
+		C_Timer.After(1, function()
+			icon:Hide()
+		end)
+	end
+end
+--]=]
+
+--end